Clean up Mixins
