--- src/main/java/me/pianopenguin471/mixins/AttackMixin.java
@@ -1,20 +1,0 @@
-package me.pianopenguin471.mixins;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.multiplayer.PlayerControllerMP;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
-import net.weavemc.loader.api.event.EventBus;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import ravenweave.client.event.impl.AttackEntityEvent;
-
-@Mixin(PlayerControllerMP.class)
-public class AttackMixin {
-    @Inject(method = "attackEntity", at = @At("HEAD"))
-    public void onAttackEntity(EntityPlayer playerIn, Entity targetEntity, CallbackInfo ci) {
-        EventBus.callEvent(new AttackEntityEvent(Minecraft.getMinecraft(), targetEntity, playerIn));
-    }
-}
