+++ src/main/java/me/pianopenguin471/mixins/MinecraftMixin.java
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import ravenweave.client.event.impl.GameLoopEvent;
 
 @Mixin(priority = 1005, value = Minecraft.class)
-public class MixinMinecraft {
+public class MinecraftMixin {
 
     @Inject(method = "runTick", at = @At("HEAD"))
     public void onTick(CallbackInfo ci) {
