+++ src/main/java/me/pianopenguin471/mixins/NetworkManagerMixin.java
@@ -12,7 +12,7 @@ import ravenweave.client.event.ext.EventDirection;
 import ravenweave.client.event.impl.PacketEvent;
 
 @Mixin(priority = 995, value = NetworkManager.class)
-public class MixinNetworkManager {
+public class NetworkManagerMixin {
 
     @Inject(method = "sendPacket(Lnet/minecraft/network/Packet;)V", at = @At("HEAD"), cancellable = true)
     public void sendPacket(Packet p_sendPacket_1_, CallbackInfo ci) {
