+++ src/main/java/me/pianopenguin471/mixins/RendererLivingEntityMixin.java
@@ -5,11 +5,13 @@ import net.minecraft.client.renderer.entity.Render;
 import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.client.renderer.entity.RendererLivingEntity;
 import net.minecraft.entity.EntityLivingBase;
+import net.weavemc.loader.api.event.EventBus;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import ravenweave.client.event.impl.RenderLabelEvent;
 
 @Mixin(RendererLivingEntity.class)
 public abstract class RendererLivingEntityMixin<T extends EntityLivingBase> extends Render<T> {
@@ -23,15 +25,11 @@ public abstract class RendererLivingEntityMixin<T extends EntityLivingBase> exte
         return input + 90;
     }
 
-    @Inject(method = "doRender(Lnet/minecraft/entity/EntityLivingBase;DDDFF)V", at=@At("HEAD"))
-    public void onDoRender(T entityIn, double var2, double var4, double var6, float yaw, float partialTicks, CallbackInfo ci) {
-        if (!(entityIn instanceof EntityPlayerSP)) return;
-        /*System.out.println(
-                "var2: " + var2 +
-                ", var4: " + var4 +
-                ", var6: " + var6 +
-                ", yaw: " + yaw +
-                ", partialTicks: " + partialTicks
-        );*/
+    @Inject(method = "renderName(Lnet/minecraft/entity/EntityLivingBase;DDD)V", at = @At("HEAD"), cancellable = true)
+    private void onRenderLabel(EntityLivingBase entity, double x, double y, double z, CallbackInfo ci) {
+        RenderLabelEvent e = new RenderLabelEvent(entity, x, y, z);
+        EventBus.callEvent(e);
+        if (e.isCancelled())
+            ci.cancel();
     }
 }
